// src/types/order.ts
import { CartItem, ShippingAddress } from "./Cart";
import { User } from "./User";

export type Order = {
  _id?: string;
  orderItems: CartItem[];
  shippingAddress?: ShippingAddress;
  paymentMethod?: string;
  user?: User | string;
  createdAt?: string;
  isPaid?: boolean;
  paidAt?: string;
  isDelivered?: boolean;
  deliveredAt?: string;
  itemsPrice: number;
  shippingPrice: number;
  taxPrice: number;
  totalPrice: number;
};